assignment/GENAI/policy_claims_copilot_project/requirements.txt
langchain
langchain-community
langchain-text-splitters
python-dotenv
langchain-openai
openai
faiss-cpu
pypdf
fastapi
uvicorn
sentence-transformers